LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/audioworkletglobalscope-creation-time.https-expected.txt


FAIL currentTime vs currentFrames in Audio Worklet global scope assert_equals: Expected workletFrame (0) to be equal to contextTime * sampleRate (41728) expected 0 but got 41728
